Move EventBus::Testing to separate file (not required by default)
