+++ lib/railjet/event_bus.rb
@@ -31,23 +31,5 @@ module Railjet
     private
 
     attr_reader :bus
-
-    module Testing
-      begin
-        require "wisper/testing"
-      rescue LoadError
-        # It's not there in staging and production env
-      end
-
-      class << self
-        delegate :adapter, to: EventBus
-        delegate :clear,   to: :adapter
-        delegate :inline,  to: :testing
-
-        def testing
-          adapter::Testing
-        end
-      end
-    end
   end
 end
