+++ spec/spec_helper.rb
@@ -3,3 +3,4 @@ $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require "pry"
 
 require 'railjet'
+require 'railjet/event_bus/testing'
